requirements.txt
flask
flask-cors
gunicorn
flask-sqlalchemy
psycopg2-binary
alembic
Flask-JWT-Extended
passlib
flasgger
